+++ test/integration/routing_test.rb
@@ -18,7 +18,6 @@ class RoutingTest < Redmine::RoutingTest
   end
 
   def test_auto_completes
-    should_route 'GET /auto_completes/fontawesome' => 'auto_completes#fontawesome'
     should_route 'GET /auto_completes/issue_assignee' => 'auto_completes#issue_assignee'
     should_route 'GET /auto_completes/assignee' => 'auto_completes#assignee'
     should_route 'GET /auto_completes/authors' => 'auto_completes#authors'