+++ test/unit/helpers/additionals_icons_helper_test.rb
@@ -1,0 +1,122 @@
+# frozen_string_literal: true
+
+require File.expand_path '../../../test_helper', __FILE__
+
+class AdditionalsIconsHelperTest < Additionals::HelperTest
+  include AdditionalsIconsHelper
+
+  def setup
+    AdditionalsIcon.reset!
+  end
+
+  def test_additionals_icon_renders_legacy_value_from_main_sprite
+    html = additionals_icon 'fas_car'
+
+    assert_match(/<svg\b/, html)
+    assert_match %r{#icon--car}, html
+    assert_no_match(/icons_custom/, html)
+  end
+
+  def test_additionals_icon_renders_tabler_name
+    html = additionals_icon 'brand-gitlab'
+
+    assert_match %r{#icon--brand-gitlab}, html
+  end
+
+  def test_additionals_icon_renders_custom_brand_from_custom_sprite
+    html = additionals_icon 'matomo'
+
+    assert_match(/icons_custom.*#icon--matomo/m, html)
+  end
+
+  def test_additionals_icon_renders_nothing_for_blank
+    assert_equal '', additionals_icon('')
+    assert_equal '', additionals_icon(nil)
+  end
+
+  def test_additionals_icon_translates_semantic_gap
+    html = additionals_icon 'fas_wrench'
+
+    assert_match %r{#icon--tool}, html
+  end
+
+  def test_additionals_icon_renders_bare_fontawesome_name
+    # bare FontAwesome names from the {{fa}}/{{tabler}} wiki macros
+    assert_match %r{#icon--file}, additionals_icon('file-alt')
+    assert_match %r{#icon--tool}, additionals_icon('wrench')
+    assert_match %r{#icon--list-numbers}, additionals_icon('list-ol')
+  end
+
+  def test_additionals_icon_applies_size_class
+    html = additionals_icon 'list-ol', size: 40, css_class: 'additionals-macro-icon'
+
+    assert_match %r{#icon--list-numbers}, html
+    assert_match(/class="s40 icon-svg additionals-macro-icon"/, html)
+  end
+
+  def test_icon_select_options_carry_main_sprite_href
+    car = send(:additionals_icon_select_options, nil).find { |option| option[1] == 'car' }
+
+    assert_not_nil car
+    assert_match %r{/additionals/icons-\w+\.svg#icon--car\z}, car[2]['data-href']
+    assert_no_match(/icons_custom/, car[2]['data-href'])
+  end
+
+  def test_icon_select_options_carry_custom_sprite_href
+    matomo = send(:additionals_icon_select_options, nil).find { |option| option[1] == 'matomo' }
+
+    assert_not_nil matomo
+    assert_match(/icons_custom.*#icon--matomo/, matomo[2]['data-href'])
+  end
+
+  def test_icon_select_options_include_known_selected_outside_subset
+    values = send(:additionals_icon_select_options, 'circle-check').pluck(1)
+
+    assert_includes values, 'circle-check'
+  end
+
+  def test_additionals_icon_renders_post_text_after_icon
+    html = additionals_icon 'car', post_text: 'My Car'
+
+    assert_match %r{#icon--car.*My Car}m, html
+  end
+
+  def test_additionals_icon_renders_pre_text_before_icon
+    html = additionals_icon 'car', pre_text: 'Label'
+
+    assert_match %r{Label.*#icon--car}m, html
+  end
+
+  def test_additionals_icon_wraps_title_in_span
+    html = additionals_icon 'car', title: 'Tooltip'
+
+    assert_match %r{<span title="Tooltip">.*#icon--car.*</span>}m, html
+  end
+
+  def test_additionals_icon_accepts_class_alias_for_css_class
+    html = additionals_icon 'car', class: 'my-icon'
+
+    assert_match(/my-icon/, html)
+  end
+
+  def test_additionals_icon_select_tag_renders_select_with_options
+    html = additionals_icon_select_tag 'menu_symbol', 'car', class: 'select2-icon-field menu'
+
+    assert_match %r{<select[^>]*name="menu_symbol"}, html
+    assert_match(/select2-icon-field menu/, html)
+    assert_match %r{<option[^>]*selected="selected"[^>]*value="car"}, html
+  end
+
+  def test_additionals_icon_select_tag_resolves_legacy_selected
+    html = additionals_icon_select_tag 'menu_symbol', 'fas_car'
+
+    assert_match %r{<option[^>]*selected="selected"[^>]*value="car"}, html
+  end
+
+  def test_additionals_icon_loader_renders_select2_init
+    js = additionals_icon_loader field_class: 'select2-menu-icon-field'
+
+    assert_match(/select2-menu-icon-field/, js)
+    assert_match(/formatIconOption/, js)
+  end
+end